requirements.txt
Django==4.0
djangorestframework==3.13.1
drf-spectacular==0.22.1
gunicorn==20.1.0
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
psycopg2-binary==2.9.5
django-filter